refactor(engine): extract finish_file + flush_workers into parallel_apply/drain.rs (SPL-38.e)
